feat(Components): Don't auto-inject CSS variables and fonts
BREAKING CHANGE: In order to improve performance when using the Components
library with other stylesheets, or bundling dependencies, we now no longer
automatically inject webfonts or CSS variables.
To reinstate the previous behaviour, you can use the provided utility functions,
`loadFonts` and `injectThemeVariables`.
