+++ packages/components/src/index.ts
@@ -4,3 +4,5 @@ export type { IconNames } from './components/icon/iconNames'
 export * as Toast from './components/toast/toastController'
 export type { Colors, ThemeColors } from './types'
 export * as FileFormatUtils from './components/editor/languageUtils'
+export { loadFonts } from './utils/fonts'
+export { injectThemeVariables } from './utils/variables'
